Mod compile script to move all compiled files to bin

+++ compile.sh
@@ -5,5 +5,19 @@
 ########################################
 
 # Update this file with instructions on how to compile your code
-ghc -dynamic src/solver.hs
-mv src/solver bin/
+ghc src/solver.hs
+
+mkdir -p bin/
+if test -f "src/solver"; then
+	mv src/solver bin/
+fi
+if test -f "src/solver.exe"; then
+	mv src/solver.exe bin/
+fi
+if test -f "src/solver.o"; then
+	mv src/solver.o bin/
+fi
+if test -f "src/solver.hi"; then
+	mv src/solver.hi bin/
+fi
+
